+++ src/main/java/com/stageEte/evaluation/repository/EvaluationRepository.java
@@ -3,5 +3,5 @@ package com.stageEte.evaluation.repository;
 import com.stageEte.evaluation.model.Evaluation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EvaluationRepository extends JpaRepository<Evaluation, Integer> {
+public interface EvaluationRepository extends JpaRepository<Evaluation, Long> {
 }
