+++ src/main/java/com/stageEte/evaluation/repository/SkillsRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SkillsRepository extends JpaRepository<Skills,Integer> {
-}
+public interface SkillsRepository extends JpaRepository<Skills,Long> { }
